requirements.txt
mediapipe==0.10.0
opencv-python==4.8.0.74
matplotlib==3.8.1
numpy==1.25.0
Pillow==10.0.1
